Refactor to remove constants dir
