--- src/state/ducks/scidata/constants/schemaList.js
@@ -1,10 +1,0 @@
-export default [
-    'dataset',
-    'measurement',
-    'methodology',
-    'parameter',
-    'resource',
-    'scidata',
-    'unit', 
-    'value',
-]
